+++ januscloud/proxy/plugin/videoroom.py
@@ -12,10 +12,10 @@ from januscloud.common.error import JanusCloudError, JANUS_ERROR_UNKNOWN_REQUEST
 from januscloud.common.schema import Schema, Optional, DoNotCare, \
     Use, IntVal, Default, SchemaError, BoolVal, StrRe, ListVal, Or, STRING, \
     FloatVal, AutoDel, StrVal, EnumVal
-from januscloud.proxy.core import backend_handle
-from januscloud.proxy.core.backend_session import get_backend_session
-from januscloud.proxy.core.plugin_base import PluginBase
-from januscloud.proxy.core.frontend_handle_base import FrontendHandleBase, JANUS_PLUGIN_OK_WAIT, JANUS_PLUGIN_OK
+from januscloud.core import backend_handle
+from januscloud.core.backend_session import get_backend_session
+from januscloud.core.plugin_base import PluginBase
+from januscloud.core.frontend_handle_base import FrontendHandleBase, JANUS_PLUGIN_OK_WAIT, JANUS_PLUGIN_OK
 import os.path
 from januscloud.common.confparser import parse as parse_config
 import time